GFC-705 tests now run successfully and also the services load properly.

+++ test/uk/gov/hmrc/gform/dms/DmsSubmissionControllerSpec.scala
@@ -71,7 +71,7 @@ class DmsSubmissionControllerSpec extends Spec {
 
     (mockFileUpload
       .createEnvelope(_: FormTemplateId, _: LocalDateTime)(_: HeaderCarrier))
-      .expects(*, LocalDateTime.now(clock).plusDays(mockConfig.formExpiryDays), *)
+      .expects(*, LocalDateTime.now(clock).plusDays(mockConfig.formExpiryDays.toLong), *)
       .returning(Future.successful(EnvelopeId("some-envelope-id")))
 
     (mockPdfGenerator
